fix: Resolve circular dependency
