+++ src/firebase-app.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { Credential, GoogleOAuthAccessToken } from './credential/credential';
+import { Credential, GoogleOAuthAccessToken } from './credential/google-oauth-access-token';
 import { getApplicationDefault } from './credential/credential-internal';
 import * as validator from './utils/validator';
 import { deepCopy, deepExtend } from './utils/deep-copy';
